# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API documentation",
        "description": "API for Waitlists",
        "contact": {
            "url": "https://github.com/LinijaVP"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Api server"
        }
    ],
    "paths": {
...
        "/wantlist": {
            "get": {
                "tags": [
                    "Wantlists"
                ],
                "summary": "Get list of wantlists",
                "description": "Returns list of wantlists",
                "operationId": "getWantlist",
                "parameters": [
                    {
                        "name": "customerId[eq]",
                        "in": "query",
                        "description": "Optional parameter to search by customer Id",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WantlistResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Wantlists"
                ],
                "summary": "Creates a wantlist",
                "description": "Create a new wantlist",
                "operationId": "postWantlist",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreWantlistRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WantlistResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/wantlist/{id}": {
            "get": {
                "tags": [
                    "Wantlists"
                ],
                "summary": "Get the wantlist with given id",
                "description": "Returns the wantlist with the given id",
                "operationId": "getWantlistById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Wantlist id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WantlistResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Wantlists"
                ],
                "summary": "Edit a wantlist",
                "description": "Edit a wantlist",
                "operationId": "editWantlist",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Wantlist id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateWantlistRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WantlistResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Wantlists"
                ],
                "summary": "Delete a wantlist",
                "description": "Delete a wantlist",
                "operationId": "deleteWantlist",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Wantlist id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Wantlists"
                ],
                "summary": "Patch a wantlist",
                "description": "Patch a wantlist",
                "operationId": "patchWantlist",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Wantlist id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PatchWantlistRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WantlistResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreWantlistRequest": {
                "required": [
                    "customerId",
                    "price",
                    "item",
                    "status",
                    "boughtDate"
                ],
                "properties": {
                    "customerId": {
                        "description": "Get the validation rules that apply to the request.",
                        "type": "integer",
                        "example": 105
                    },
                    "price": {
                        "type": "integer",
                        "example": 100
                    },
                    "item": {
                        "type": "string",
                        "example": "itemName"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "W",
                            "B"
                        ],
                        "example": "W"
                    },
                    "boughtDate": {
                        "type": "[date_format:Y-m-d H:i:s|nullable]",
                        "example": null
                    }
                },
                "type": "object"
            },
...
            "UpdateWantlistRequest": {
                "required": [
                    "price",
                    "item",
                    "status",
                    "boughtDate"
                ],
                "properties": {
                    "price": {
                        "description": "Get the validation rules that apply to the request.",
                        "type": "integer",
                        "example": 100
                    },
                    "item": {
                        "type": "string",
                        "example": "itemName"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "W",
                            "B"
                        ],
                        "example": "W"
                    },
                    "boughtDate": {
                        "type": "[date_format:Y-m-d H:i:s|nullable]",
                        "example": null
                    }
                },
                "type": "object"
            },
            "PatchWantlistRequest": {
                "properties": {
                    "price": {
                        "type": "integer",
                        "example": 100
                    },
                    "item": {
                        "type": "string",
                        "example": "itemName"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "W",
                            "B"
                        ],
                        "example": "W"
                    },
                    "boughtDate": {
                        "type": "[date_format:Y-m-d H:i:s|nullable]",
                        "example": null
                    }
                },
                "type": "object"
            },
...
            "WantlistResource": {
                "properties": {
                    "id": {
                        "description": "Transform the resource into an array.",
                        "type": "integer"
                    },
                    "customerId": {
                        "type": "integer"
                    },
                    "price": {
                        "type": "integer",
                        "example": 10
                    },
                    "item": {
                        "type": "string",
                        "example": "itemName"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "W",
                            "B"
                        ],
                        "example": "W"
                    },
                    "boughtDate": {
                        "type": "[date_format:Y-m-d H:i:s|nullable]",
                        "example": null
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "name": "sanctum",
                "in": "header",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}